+++ pyfakefs/fake_filesystem.py
@@ -114,7 +114,7 @@ from pyfakefs.helpers import make_string_path, text_type
 
 __pychecker__ = 'no-reimportself'
 
-__version__ = '3.6'
+__version__ = '3.5.2'
 
 PERM_READ = 0o400  # Read permission bit.
 PERM_WRITE = 0o200  # Write permission bit.