+++ CHANGES.md
@@ -8,8 +8,7 @@ The release versions are PyPi releases.
     `st_uid` and `st_gid` in new files ([#449](../../issues/449))
 
 #### Fixes
-  * fixed using `modules_to_patch` with modules without `dir()`
-    ([#450](../../issues/450))
+  * fixed using `modules_to_patch` ([#450](../../issues/450))
   * fixed recursion error on unpickling the fake file system 
     ([#445](../../issues/445))
   * allow trailing path in `add_real_directory` ([#446](../../issues/446))